Added file
This file maps out the Earth's electric field for reference in other MMS projects I am working on. It works by using multiple vectors in cartesian coordinates that are then manually converted to spherical and made in reference to the sun. If run in tkinter it can be interactive, otherwise it will not be.